==> 2_project_results_stats.ipynb <==
Create statstical results for the mined results. Specifically how many projects contain READMEs, RELEASES, both or none.

import json
import os
import plotly.express as px
import pandas as pd
import plotly.graph_objects as go

languages = ["go", "java", "python", "ruby"]
project_details_folder = "project_details"
readmes_folder = "readmes"
releases_folder = "releases"

all_language_stats = []

for language in languages:

    total_projects = 0
    total_releases = 0
    total_readmes = 0

    language_readme_folder = f'{readmes_folder}/{language}'
    language_releases_folder = f'{releases_folder}/{language}'
 
    f = open(f'{project_details_folder}/filtered/{language}_projects_filtered.json')
    projects_data = json.load(f)

    for project in projects_data:

        projectId = project["ProjectID"]
        projectOwner = projectId.split('/')[0]
        projectName = projectId.split('/')[1]
        #readme = False
        #release = False

        # Folders to store the projects readme & releases
        readme_file = f"{language_readme_folder}/{projectOwner}_{projectName}_readme.md"
        release_file = f"{language_releases_folder}/{projectOwner}_{projectName}_releases.jsonl"

        total_projects+=1
        if  os.path.exists(release_file):
            total_releases+=1
        if os.path.exists(readme_file):
            total_readmes+=1

    # language_data = dict(
    #     character=["Java", "READMEs"],
    #     parent=["", "Java",],
    #     value=[total_projects, total_readmes]
    # )

    language_stats = {"language": language, "total_projects": total_projects, "total_readmes": total_readmes, "total_releases": total_releases, "percetage_readmes": round(((total_readmes/total_projects)*100),2), "percentage_releases": round(((total_releases/total_projects)*100),2)  }
    all_language_stats.append(language_stats)
        


stats_df = pd.DataFrame(all_language_stats, columns=all_language_stats[0].keys())
display(stats_df)


fig = px.histogram(stats_df, x="language", y=["percetage_readmes"],
             color='language', 
             text_auto=True,
             height=400, width=1000)
fig2 = px.histogram(stats_df, x="language", y=["percentage_releases"],
color='language', 
text_auto=True,
height=400, width=1000)
fig.show()
fig2.show()
