==> 5_repo_readme_processor.ipynb <==
import pandas as pd
import os
import markdown
from bs4 import BeautifulSoup
import re
import csv
import nltk
import nltk.data
nltk.download("popular")
from nltk import pos_tag
from nltk.tokenize import TweetTokenizer
from nltk import pos_tag_sents
from nltk.corpus import stopwords
from nltk import pos_tag
from nltk.corpus import wordnet as wn
from nltk.stem import WordNetLemmatizer
from nltk.sentiment.util import mark_negation
import emoji
import contractions
import json
import glob

def get_lemmatizer_pos(pos):
    """Retrieves the POS-tag in a WordNet-compatible form. It returns a noun tag by default if the word is not
    an adjective, verb, or adverb.

    Args:
        pos: NTLK POS-tag.

    Returns:
        A WordNet-compatible POS-tag string.
    """
    pos_start = pos[0]  # Takes the first letter to simplify the POS tag
    if pos_start == "J":
        return wn.ADJ
    elif pos_start == "V":
        return wn.VERB
    elif pos_start == "R":
        return wn.ADV
    else:
        return wn.NOUN

def tokenize_normalize(
    context,
    sentence_tokenizer=nltk.data.load("tokenizers/punkt/english.pickle"),
    lemmatizer=WordNetLemmatizer(),
    tokenizer=TweetTokenizer(preserve_case=False), # This tokenizer can handle URLs better
    tokenize_numbers=False,
    tokenize_urls=False,
    deemojize=True,
    remove_punct=True,
    handle_negation=True
):
    stop_words = stopwords.words("english")

    punctuations = "!\"“”#$%&'‘’()*+,-./:;<=>?@[\]^_`{|}~‍，（）。•、–—？！：；［］【 】"

    sentences = sentence_tokenizer.tokenize(context)

    sentences_tokens = [tokenizer.tokenize(sentence) for sentence in sentences]

    tokens_pos = [
        pos_tag
        for pos_tags in pos_tag_sents(sentences_tokens)
        for pos_tag in pos_tags
    ]

    lemmas = [
        lemmatizer.lemmatize(token[0], pos=get_lemmatizer_pos(token[1]))
        for token in tokens_pos
    ]
    
    if handle_negation:
        lemmas = mark_negation(lemmas)

    filtered_lemmas = []
    last_filtered_lemma_index = None
    last_filtered_lemma = None
    for lemma_index, lemma in enumerate(lemmas):

        # Removing variation selectors such as hair/skin color and gender for
        # emojis since they cause noise and tokenization problems:
        if re.sub("[\\uFE00-\\uFE0F♂♀‍]+", "", lemma) == "":
            continue

        # Filters stop words (considers negations):
        if lemma.replace("_NEG", "") in stop_words:
            continue

        # Filters the lemma by searching for "https://," "http://," or "www." using regular expression. If one
        # of them exists, they are not retrieved. Regular expression may seem daunting at first. It is not
        # mandatory, but you can check tutorials like this: https://regexone.com/lesson/introduction_abcs
        if re.search("(https?:\/\/)|(www\.)", lemma):
            if tokenize_urls:
                lemma = ""
            else:
                continue

        # Deemojizes emojis using emeji package (considers negations):
        if deemojize:
            lemma = emoji.demojize(lemma)

        # Filters punctuation (considers negations):
        if (
            remove_punct
            and lemma.replace("_NEG", "").translate(
                lemma.maketrans("", "", punctuations)
            )
            == ""
        ):
            continue

        # Tries to convert a number from string to float while also handling commas and percentage signs. If
        # the token is a number, it is transformed to "" token or not retrieved. If not, it silently
        # ignores the exception and continues.
        try:
            float(lemma.replace(",", "").replace("%", ""))
            if tokenize_numbers:
                lemma = ""
            else:
                continue
        except:
            pass

        filtered_lemmas.append(lemma)
    
    return filtered_lemmas

Processes the paragraphs that come before the first H2 title in the readme file for each repository:

languages = ["Go", "Java", "Ruby", "Python"]
# languages = ["Java", "Python"]
readme_folder = "readmes"

html_tags = re.compile("<.*?>|&([a-z0-9]+|#[0-9]{1,6}|#x[0-9a-f]{1,6});")
chinese_chars = re.compile(u"[⺀-⺙⺛-⻳⼀-⿕々〇〡-〩〸-〺〻㐀-䶵一-鿃豈-鶴侮-頻並-龎]", re.UNICODE)

for language in languages:
    
    print(f"Processing {language} repositories...")
    
    processed = set()
    
    with open(f"processed_readmes_{language.lower()}.csv", "w", newline="", encoding="UTF-8") as csv_file:
        csv_writer = csv.writer(csv_file, delimiter=";", quotechar="'", quoting=csv.QUOTE_MINIMAL)
        csv_writer.writerow(["OWNER", "NAME", "CONTEXTS"])
        
        language_repos = glob.glob(f"{readme_folder}/{language.lower()}/*.md")
    
        improper_readme_count = 0

        for file in language_repos:
            owner, repo_name = os.path.basename(file).split("_", 1)
            
            repo_name = repo_name.replace("_readme.md", "")
            
            if f"{owner}_{repo_name}" in processed:
                print(f"WARNING: {owner}_{repo_name} was already processed. Skipping this repository...")
                continue
            else:
                processed.add(f"{owner}_{repo_name}")

            with open(file, encoding="UTF-8") as f:
                readme = f.read()
                
            # Not all readme files are perfectly parseable. Some extra steps were needed both before and after parsing the file.
            if "```" in readme:
                readme = readme.split("```")
                if readme[0] == "":
                    readme = "\n".join(readme[1::2])
                else:
                    readme = "\n".join(readme[0::2])

            readme = BeautifulSoup(markdown.markdown(readme), "html.parser")
            for img in readme.select("img"):
                img.extract()

            extracted_text = []
            improper_formatting = True if all([True if e.name != "h1" and e.name != "h2" else False for e in readme for e in readme]) else False
            if improper_formatting:
                improper_readme_count += 1
                continue
            
            e_counter = 0
            
            for e_id, e in enumerate(readme):
                # if e.name == "h2":
                #     break
                
                # if e_id == 5:
                #     break
                
                if e_counter == 1:
                    break
                
                if e.name == "p": # Clear-text paragraphs are only retrieved if they satisfy certain conditions
                # if 1:
                    # e_counter += 1
                    e = html_tags.sub("", e.text)
                    e = chinese_chars.sub("", e) # Removing Chinese characters
                    e = contractions.fix(e) # Expanding contractions
                    # if language.lower() not in e.lower():
                    #     continue
                    # if "library" not in e.lower() and "package" not in e.lower():
                    #     continue
                    # if ("library" not in e.lower() and "package" not in e.lower()) or language.lower() not in e.lower():
                    #     continue
                    if len(e.strip()) > 3:
                        extracted = " ".join([e for e in e.split("\n") if not e.startswith(".. image::") if not e.startswith("image:") if not e.startswith("  :") if not ":target:" in e if not ":alt:" in e])
                        extracted = tokenize_normalize(extracted.strip().encode("ascii", "ignore").decode("utf-8")) # non-ASCII characters are removed
                        if len(set(extracted)) > 2:
                            extracted_text.append(extracted)
                            e_counter += 1

                # if e.name == "h2":
                #     e = html_tags.sub("", e.text)
                #     e = chinese_chars.sub("", e) # Removing Chinese characters
                #     e = contractions.fix(e) # Expanding contractions
                #     if len(e.strip()) > 3:
                #         extracted = " ".join([e for e in e.split("\n") if not e.startswith(".. image::") if not e.startswith("image:") if not e.startswith("  :")])
                #         extracted = tokenize_normalize(extracted.strip().encode("ascii", "ignore").decode("utf-8")) # non-ASCII characters are removed
                #         if len(set(extracted)) > 2:
                #             extracted_text.append(extracted)
            
            if extracted_text:
                csv_writer.writerow([owner, repo_name, json.dumps(extracted_text, ensure_ascii=False)])
                
        print(f"Ignored improper readme files for {language}: {improper_readme_count}/{len(language_repos)}")
